+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `AccountCertificateEndpoint` and `Certificate` resource
 - Add `Label` resource and `LabelsEndpoint`
 - Add `EnvelopeLabelsEndpoint`
+- Add `EnvelopeRecipient.bankIdScopes`
 
 ## [1.6.0] - 2021-11-26
 ### Added