+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeLabelsEndpoint`
 - Add `EnvelopeRecipient.bankIdScopes`
 - Add `AccountCertificatesEndpoint.enable` and `AccountCertificatesEndpoint.disable`
+- Add `EnvelopeTemplateLabelsEndpoint`
 
 ### Changed
 - Replace fields `authenticationMethod`, `authenticationPlace` and `authenticateOnDownload` with `authenticationOnOpen`, `authenticationOnSignature` and `authenticationOnDownload`