+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [Unreleased]
+## [1.7.0]
 ### Added
 - Add `AccountCertificateEndpoint` and `Certificate` resource
 - Add `Label` resource and `LabelsEndpoint`